Backend/requirements_x86.txt
absl-py==0.10.0
aiohttp==3.6.2
aniso8601==8.0.0
asgiref==3.3.1
astunparse==1.6.3
async-timeout==3.0.1
attr==0.3.1
attrs==20.2.0
bert-for-tf2==0.14.7
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coremltools==4.0
Django==3.1.4
firebase-admin==4.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
flatbuffers==1.12
gast==0.3.3
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.2
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
idna==2.10
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.6
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
params-flow==0.8.2
Pillow==8.0.1
protobuf==3.13.0
py-params==0.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.19.2
scipy==1.5.2
sentencepiece==0.1.94
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
sympy==1.6.2
syncer==1.3.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.10.0
tensorflow-text==2.4.2
termcolor==1.1.0
tf-sentencepiece==0.1.90
threadpoolctl==2.1.0
tqdm==4.50.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
